editoast: views: move `study` in `operational_studies::hierarchy`
