+++ editoast/src/views/operational_studies/hierarchy.rs
@@ -1,1 +1,2 @@
 pub mod scenario;
+pub mod study;
