+++ editoast/src/views/operational_studies/hierarchy/scenario.rs
@@ -18,6 +18,7 @@ use thiserror::Error;
 use utoipa::IntoParams;
 use utoipa::ToSchema;
 
+use super::study::StudyError;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::views::operational_studies::OperationalStudiesOrderingParam;
@@ -25,7 +26,6 @@ use crate::views::pagination::PaginatedList as _;
 use crate::views::pagination::PaginationQueryParams;
 use crate::views::pagination::PaginationStats;
 use crate::views::project::ProjectError;
-use crate::views::study::StudyError;
 use editoast_models::Infra;
 use editoast_models::project::Project;
 use editoast_models::scenario::Scenario;
